DagsterInstance domain refactoring plans and extract runs code (#31510)
## Summary & Motivation

This PR refactors Dagster's run implementation from plain functions to a domain-based architecture. We extract 9 run-related methods from the monolithic `DagsterInstance` class into a dedicated `RunDomain` class, establishing a proven pattern for future domain extractions.

The new architecture introduces a `RunDomain` class that holds a reference to `DagsterInstance` and calls its methods directly. `DagsterInstance` uses `@cached_property` for lazy initialization and delegates run operations to the domain.

```python
class RunDomain:
    def __init__(self, instance: "DagsterInstance") -> None:
        self._instance = instance
    
    def create_run(self, *, job_name: str, ...):
        # Direct calls to self._instance methods
        return self._instance.run_storage.add_run(dagster_run)

class DagsterInstance:
    @cached_property
    def _run_domain(self):
        return RunDomain(self)
    
    def create_run(self, *, job_name: str, ...):
        return self._run_domain.create_run(job_name=job_name, ...)
```

This reduces `DagsterInstance` by 853 lines while maintaining perfect backwards compatibility. All existing APIs work unchanged and subclass overrides remain functional.

## How I Tested These Changes

Existing test suite continues to pass with zero modifications required.
EOF < /dev/null
